feat: Data flow successfully integrated
Components have been successfully linked between themselves and fed with data from the API. Books are correctly displayed depending on the shelf they are supposed to be on. The "Changer" button (which allows to change the shelf of the book) is working, re-rendering the components without reloading the page, updating the API and the state of the books.

- Added a dependency to generate random unique keys to components in arrays.