feat: fixing the rendering of components
Successfully fixed the render of components: now, every new book added through the search page will trigger a re-render of the app to display the new book on its proper shelf. The book button shelf status changer is also immediately updated whether on the search page or the shelves page.